+++ src/ts/Ads/Ads.ts
@@ -32,7 +32,7 @@ import { AdsConfigurationParser } from 'Ads/Parsers/AdsConfigurationParser';
 import { CustomFeatures } from 'Ads/Utilities/CustomFeatures';
 import { GameSessionCounters } from 'Ads/Utilities/GameSessionCounters';
 import { IosUtils } from 'Ads/Utilities/IosUtils';
-import { ChinaMetric, ProgrammaticTrackingError, MiscellaneousMetric, LoadMetric, ProgrammaticTrackingService, InitializationMetric } from 'Ads/Utilities/ProgrammaticTrackingService';
+import { ChinaMetric, ErrorMetric, MiscellaneousMetric, LoadMetric, SDKMetrics, InitializationMetric } from 'Ads/Utilities/SDKMetrics';
 import { SdkStats } from 'Ads/Utilities/SdkStats';
 import { SessionDiagnostics } from 'Ads/Utilities/SessionDiagnostics';
 import { InterstitialWebPlayerContainer } from 'Ads/Utilities/WebPlayer/InterstitialWebPlayerContainer';
@@ -84,12 +84,11 @@ import { Analytics } from 'Analytics/Analytics';
 import { PrivacySDK } from 'Privacy/PrivacySDK';
 import { PrivacyParser } from 'Privacy/Parsers/PrivacyParser';
 import { Promises } from 'Core/Utilities/Promises';
-import { LoadExperiment, LoadRefreshV4, AdmobAdapterV1 } from 'Core/Models/ABGroup';
+import { LoadExperiment, LoadRefreshV4 } from 'Core/Models/ABGroup';
 import { PerPlacementLoadManagerV4 } from 'Ads/Managers/PerPlacementLoadManagerV4';
 import { PrivacyMetrics } from 'Privacy/PrivacyMetrics';
 import { PerPlacementLoadAdapter } from 'Ads/Managers/PerPlacementLoadAdapter';
 import { PrivacyDataRequestHelper } from 'Privacy/PrivacyDataRequestHelper';
-import { AdmobAdapterManager } from 'Ads/Managers/AdmobAdapterManager';
 import { MediationMetaData } from 'Core/Models/MetaData/MediationMetaData';
 import { MediationLoadTrackingManager } from 'Ads/Managers/MediationLoadTrackingManager';
 
@@ -113,7 +112,6 @@ export class Ads implements IAds {
     public AssetManager: AssetManager;
     public CampaignManager: CampaignManager;
     public RefreshManager: RefreshManager;
-    public AdmobAdapterManager: AdmobAdapterManager;
     public MediationLoadTrackingManager: MediationLoadTrackingManager;
 
     private static _forcedConsentUnit: boolean = false;
@@ -184,8 +182,6 @@ export class Ads implements IAds {
             GameSessionCounters.init();
             Diagnostics.setAbGroup(this._core.Config.getAbGroup());
             return this.setupTestEnvironment();
-        }).then(() => {
-            return this.configureMediationManager();
         }).then(() => {
             return this.Analytics.initialize();
         }).then((gameSessionId: number) => {
@@ -311,26 +307,6 @@ export class Ads implements IAds {
         }
     }
 
-    private configureMediationManager(): Promise<void> {
-        const allowedByAbTest = AdmobAdapterV1.isValid(this._core.Config.getAbGroup());
-        const allowedByGameId = CustomFeatures.isAdmobTimeoutWhitelisted(this._core.ClientInfo.getGameId());
-
-        if (allowedByAbTest || allowedByGameId) {
-            return this._core.MetaDataManager.fetch(MediationMetaData).then((mediation) => {
-                if (mediation) {
-                    const mediationName = mediation.getName();
-                    if (mediationName === 'AdMob') {
-                        this.AdmobAdapterManager = new AdmobAdapterManager(this.Api, this._core.NativeBridge.getPlatform());
-                    }
-                }
-            }).catch(() => {
-                // ingore error
-            });
-        }
-
-        return Promise.resolve();
-    }
-
     private setupMediationTrackingManager(): Promise<void> {
         if (this._loadApiEnabled) {
 
@@ -352,15 +328,15 @@ export class Ads implements IAds {
         if (initTimestamp && performance && performance.now) {
             const webviewInitTime = performance.now() - initTimestamp;
             const tags = [
-                ProgrammaticTrackingService.createAdsSdkTag('wel', `${this._webViewEnabledLoad}`),
-                ProgrammaticTrackingService.createAdsSdkTag('lae', `${this._loadApiEnabled}`)
+                SDKMetrics.createAdsSdkTag('wel', `${this._webViewEnabledLoad}`),
+                SDKMetrics.createAdsSdkTag('lae', `${this._loadApiEnabled}`)
             ];
 
             if (this._mediationName) {
-                tags.push(ProgrammaticTrackingService.createAdsSdkTag('med', this._mediationName));
+                tags.push(SDKMetrics.createAdsSdkTag('med', this._mediationName));
             }
-            ProgrammaticTrackingService.reportTimingEventWithTags(InitializationMetric.WebviewInitialization, webviewInitTime, tags);
-            ProgrammaticTrackingService.reportTimingEventWithTags(InitializationMetric.WebviewPageLoading, initTimestamp, tags);
+            SDKMetrics.reportTimingEventWithTags(InitializationMetric.WebviewInitialization, webviewInitTime, tags);
+            SDKMetrics.reportTimingEventWithTags(InitializationMetric.WebviewPageLoading, initTimestamp, tags);
         }
     }
 
@@ -400,24 +376,21 @@ export class Ads implements IAds {
         callback(CallbackStatus.OK);
 
         if (this.isAttemptingToShowInBackground()) {
-            ProgrammaticTrackingService.reportMetricEvent(MiscellaneousMetric.CampaignAttemptedShowInBackground);
+            SDKMetrics.reportMetricEvent(MiscellaneousMetric.CampaignAttemptedShowInBackground);
             return;
         }
 
         const campaign = this.RefreshManager.getCampaign(placementId);
         if (!campaign) {
             this.showError(true, placementId, 'Campaign not found');
-            ProgrammaticTrackingService.reportMetricEvent(MiscellaneousMetric.CampaignNotFound);
+            SDKMetrics.reportMetricEvent(MiscellaneousMetric.CampaignNotFound);
             return;
         }
-
-        const contentType = campaign.getContentType();
-        const seatId = campaign.getSeatId();
 
         if (this._showing || this._showingPrivacy) {
             // do not send finish event because there will be a finish event from currently open ad unit
             this.showError(false, placementId, 'Can\'t show a new ad unit when ad unit is already open');
-            ProgrammaticTrackingService.reportErrorEvent(ProgrammaticTrackingError.AdUnitAlreadyShowing, contentType, seatId);
+            SDKMetrics.reportMetricEvent(ErrorMetric.AdUnitAlreadyShowing);
             return;
         }
 
@@ -438,7 +411,7 @@ export class Ads implements IAds {
         const placement: Placement = this.Config.getPlacement(placementId);
         if (!placement) {
             this.showError(true, placementId, 'No such placement: ' + placementId);
-            ProgrammaticTrackingService.reportErrorEvent(ProgrammaticTrackingError.PlacementWithIdDoesNotExist, contentType, seatId);
+            SDKMetrics.reportMetricEvent(ErrorMetric.PlacementWithIdDoesNotExist);
             return;
         }
 
@@ -446,7 +419,7 @@ export class Ads implements IAds {
 
         if (campaign instanceof PromoCampaign && campaign.getRequiredAssets().length === 0) {
             this.showError(false, placementId, 'No creatives found for promo campaign');
-            ProgrammaticTrackingService.reportErrorEvent(ProgrammaticTrackingError.PromoWithoutCreatives, contentType, seatId);
+            SDKMetrics.reportMetricEvent(ErrorMetric.PromoWithoutCreatives);
             return;
         }
 
@@ -460,7 +433,7 @@ export class Ads implements IAds {
                 contentType: campaign.getContentType()
             });
             SessionDiagnostics.trigger('campaign_expired', error, campaign.getSession());
-            ProgrammaticTrackingService.reportErrorEvent(ProgrammaticTrackingError.CampaignExpired, contentType, seatId);
+            SDKMetrics.reportMetricEvent(ErrorMetric.CampaignExpired);
             return;
         }
 
@@ -469,7 +442,7 @@ export class Ads implements IAds {
             // Do not remove: Removing will currently break all tracking
             campaign.setTrackingUrls(trackingUrls);
         } else {
-            ProgrammaticTrackingService.reportErrorEvent(ProgrammaticTrackingError.MissingTrackingUrlsOnShow, contentType);
+            SDKMetrics.reportMetricEvent(ErrorMetric.MissingTrackingUrlsOnShow);
         }
 
         this.showPrivacyIfNeeded(options).then(() => {
@@ -506,7 +479,7 @@ export class Ads implements IAds {
                 });
                 SessionDiagnostics.trigger('mraid_no_connection', error, campaign.getSession());
                 // If there is no connection, would this metric even be fired? If it does, then maybe we should investigate enabling this regardless of connection
-                ProgrammaticTrackingService.reportErrorEvent(ProgrammaticTrackingError.NoConnectionWhenNeeded, campaign.getContentType(), campaign.getSeatId());
+                SDKMetrics.reportMetricEvent(ErrorMetric.NoConnectionWhenNeeded);
                 return;
             }
 
@@ -517,7 +490,12 @@ export class Ads implements IAds {
             if (this.Monetization.isInitialized()) {
                 this.Monetization.PlacementContentManager.setCurrentAdUnit(placement.getId(), this._currentAdUnit);
             }
-            this._currentAdUnit.onClose.subscribe(() => this.onAdUnitClose());
+            this._currentAdUnit.onClose.subscribe(() =>  {
+                this.onAdUnitClose();
+                SDKMetrics.sendBatchedEvents();
+            });
+            this._currentAdUnit.onFinish.subscribe(() => SDKMetrics.sendBatchedEvents());
+            this._currentAdUnit.onError.subscribe(() => SDKMetrics.sendBatchedEvents());
 
             if (this._core.NativeBridge.getPlatform() === Platform.IOS && (campaign instanceof PerformanceCampaign || campaign instanceof XPromoCampaign)) {
                 if (!IosUtils.isAppSheetBroken(this._core.DeviceInfo.getOsVersion(), this._core.DeviceInfo.getModel(), orientation) && !campaign.getBypassAppSheet()) {
@@ -546,7 +524,7 @@ export class Ads implements IAds {
 
             this._currentAdUnit.show().then(() => {
                 if (this._loadApiEnabled && this._webViewEnabledLoad) {
-                    ProgrammaticTrackingService.reportMetricEvent(LoadMetric.LoadEnabledShow);
+                    SDKMetrics.reportMetricEvent(LoadMetric.LoadEnabledShow);
                 }
             });
         });
@@ -672,7 +650,7 @@ export class Ads implements IAds {
     private logChinaMetrics() {
         const isChineseUser = this._core.Config.getCountry() === 'CN';
         if (isChineseUser) {
-            ProgrammaticTrackingService.reportMetricEvent(ChinaMetric.ChineseUserInitialized);
+            SDKMetrics.reportMetricEvent(ChinaMetric.ChineseUserInitialized);
         }
         this.identifyUser(isChineseUser);
     }
@@ -681,7 +659,7 @@ export class Ads implements IAds {
         this.isUsingChineseNetworkOperator().then(isAChineseNetwork => {
             if (isAChineseNetwork) {
                 const networkMetric = isChineseUser ? ChinaMetric.ChineseUserIdentifiedCorrectlyByNetworkOperator : ChinaMetric.ChineseUserIdentifiedIncorrectlyByNetworkOperator;
-                ProgrammaticTrackingService.reportMetricEvent(networkMetric);
+                SDKMetrics.reportMetricEvent(networkMetric);
             } else {
                 const localeMetric = isChineseUser ? ChinaMetric.ChineseUserIdentifiedCorrectlyByLocale : ChinaMetric.ChineseUserIdentifiedIncorrectlyByLocale;
                 this.logChinaLocalizationOptimizations(localeMetric);
@@ -700,7 +678,7 @@ export class Ads implements IAds {
         const chineseLanguage = !!(deviceLanguage.match(/zh[-_]cn/) || deviceLanguage.match(/zh[-_]hans/) || deviceLanguage.match(/zh(((_#?hans)?(_\\D\\D)?)|((_\\D\\D)?(_#?hans)?))$/));
         const chineseTimeZone = this._core.DeviceInfo.getTimeZone() === 'GMT+08:00';
         if (chineseLanguage && chineseTimeZone) {
-            ProgrammaticTrackingService.reportMetricEvent(metric);
+            SDKMetrics.reportMetricEvent(metric);
         }
     }