+++ src/ts/Ads/Networking/MetricInstance.ts
@@ -1,8 +1,9 @@
-import { ProgrammaticTrackingError, PTSEvent, TimingEvent } from 'Ads/Utilities/ProgrammaticTrackingService';
+import { ErrorMetric, PTSEvent, TimingEvent } from 'Ads/Utilities/SDKMetrics';
 import { Platform } from 'Core/Constants/Platform';
 import { INativeResponse, RequestManager } from 'Core/Managers/RequestManager';
 import { ClientInfo } from 'Core/Models/ClientInfo';
 import { DeviceInfo } from 'Core/Models/DeviceInfo';
+import { Promises } from 'Core/Utilities/Promises';
 
 interface IPTSEvent {
     name: string;
@@ -21,7 +22,8 @@ export class MetricInstance {
     private _clientInfo: ClientInfo;
     private _deviceInfo: DeviceInfo;
     private _countryIso: string;
-    private _batchedEvents: IPTSEvent[];
+    private _batchedTimingEvents: IPTSEvent[];
+    private _batchedMetricEvents: IPTSEvent[];
     private _baseUrl: string;
 
     private _stagingBaseUrl = 'https://sdk-diagnostics.stg.mz.internal.unity3d.com/';
@@ -35,7 +37,8 @@ export class MetricInstance {
         this._clientInfo = clientInfo;
         this._deviceInfo = deviceInfo;
         this._countryIso = this.getCountryIso(country);
-        this._batchedEvents = [];
+        this._batchedTimingEvents = [];
+        this._batchedMetricEvents = [];
         this._baseUrl = this._clientInfo.getTestMode() ? this._stagingBaseUrl : this.getProductionUrl();
     }
 
@@ -48,22 +51,6 @@ export class MetricInstance {
             this.createAdsSdkTag('sdv', this._clientInfo.getSdkVersionName()),
             this.createAdsSdkTag('iso', this._countryIso),
             this.createAdsSdkTag('plt', Platform[this._platform])].concat(tags);
-    }
-
-    private createErrorTags(event: PTSEvent, adType?: string, seatId?: number): string[] {
-
-        const platform: Platform = this._platform;
-        const osVersion: string = this._deviceInfo.getOsVersion();
-        const sdkVersion: string = this._clientInfo.getSdkVersionName();
-
-        return [
-            this.createAdsSdkTag('eevt', event),
-            this.createAdsSdkTag('plt', Platform[platform]),
-            this.createAdsSdkTag('osv', osVersion),
-            this.createAdsSdkTag('sdv', sdkVersion),
-            this.createAdsSdkTag('adt', `${adType}`),
-            this.createAdsSdkTag('sid', `${seatId}`)
-        ];
     }
 
     private createData(event: PTSEvent, value: number, tags: string[]): IProgrammaticTrackingData {
@@ -109,65 +96,80 @@ export class MetricInstance {
         return `ads_sdk2_${suffix}:${tagValue}`;
     }
 
-    public reportMetricEvent(event: PTSEvent): Promise<INativeResponse> {
-        return this.reportMetricEventWithTags(event, []);
+    public reportMetricEvent(event: PTSEvent) {
+        this.reportMetricEventWithTags(event, []);
     }
 
     public reportMetricEventWithTags(event: PTSEvent, tags: string[]) {
-        const metricData = this.createData(event, 1, this.createTags(tags));
-        return this.postToDatadog(metricData, this.metricPath);
+        this.batchMetricEvent(event, 1, this.createTags(tags));
     }
 
-    public reportErrorEvent(event: PTSEvent, adType: string, seatId?: number): Promise<INativeResponse> {
-        const errorData = this.createData(event, 1, this.createErrorTags(event, adType, seatId));
-        return this.postToDatadog(errorData, this.metricPath);
-    }
-
-    public reportTimingEvent(event: TimingEvent, value: number): Promise<INativeResponse> {
+    public reportTimingEvent(event: TimingEvent, value: number) {
         // Gate Negative Values
         if (value > 0) {
-            const timingData = this.createData(event, value, this.createTags([]));
-            return this.postToDatadog(timingData, this.timingPath);
+            this.batchTimingEvent(event, value, this.createTags([]));
         } else {
-            const metricData = this.createData(ProgrammaticTrackingError.TimingValueNegative, 1, this.createTags([
+            this.batchMetricEvent(ErrorMetric.TimingValueNegative, 1, this.createTags([
                 this.createAdsSdkTag('mevt', event)
             ]));
-            return this.postToDatadog(metricData, this.metricPath);
         }
     }
 
-    public reportTimingEventWithTags(event: TimingEvent, value: number, tags: string[]): Promise<INativeResponse> {
+    public reportTimingEventWithTags(event: TimingEvent, value: number, tags: string[]) {
         if (value > 0) {
-            const timingData = this.createData(event, value, this.createTags(tags));
-            return this.postToDatadog(timingData, this.timingPath);
+            this.batchTimingEvent(event, value, this.createTags(tags));
+        } else {
+            this.batchMetricEvent(ErrorMetric.TimingValueNegative, 1, this.createTags([
+                this.createAdsSdkTag('mevt', event)
+            ]));
         }
-        // Probably don't care about logging negative
-        return Promise.resolve(<INativeResponse>{});
     }
 
-    // TODO: Extend this to all events
-    public batchEvent(metric: TimingEvent, value: number): void {
-        // Curently ignore additional negative time values
-        if (value > 0) {
-            this._batchedEvents = this._batchedEvents.concat(this.createData(metric, value, this.createTags([])).metrics);
-        }
+    public sendBatchedEvents(): Promise<void[]> {
+        const promises = [
+            this.sendBatchedMetricEvents(),
+            this.sendBatchedTimingEvents()
+        ];
+        return Promise.all(promises);
+    }
+
+    private sendBatchedMetricEvents(): Promise<void> {
+        return this.constructAndSendEvents(this._batchedMetricEvents, this.metricPath).then(() => {
+            this._batchedMetricEvents = [];
+        });
+    }
+
+    private sendBatchedTimingEvents(): Promise<void> {
+        return this.constructAndSendEvents(this._batchedTimingEvents, this.timingPath).then(() => {
+            this._batchedTimingEvents = [];
+         });
+    }
+
+    private batchTimingEvent(metric: PTSEvent, value: number, tags: string[]): void {
+        this._batchedTimingEvents = this._batchedTimingEvents.concat(this.createData(metric, value, tags).metrics);
 
         // Failsafe so we aren't storing too many events at once
-        if (this._batchedEvents.length >= 10) {
-            this.sendBatchedEvents();
+        if (this._batchedTimingEvents.length >= 30) {
+            this.sendBatchedTimingEvents();
         }
     }
 
-    public sendBatchedEvents(): Promise<void> {
-        if (this._batchedEvents.length > 0) {
+    private batchMetricEvent(metric: PTSEvent, value: number, tags: string[]): void {
+        this._batchedMetricEvents = this._batchedMetricEvents.concat(this.createData(metric, value, tags).metrics);
+
+        // Failsafe so we aren't storing too many events at once
+        if (this._batchedMetricEvents.length >= 30) {
+            this.sendBatchedMetricEvents();
+        }
+    }
+
+    private constructAndSendEvents(events: IPTSEvent[], path: string): Promise<void> {
+        if (events.length > 0) {
             const data = {
-                metrics: this._batchedEvents
+                metrics: events
             };
-            return this.postToDatadog(data, this.timingPath).then(() => {
-                this._batchedEvents = [];
-            });
+            return Promises.voidResult(this.postToDatadog(data, path));
         }
         return Promise.resolve();
     }
-
 }