+++ src/ts/Ads/Utilities/CustomFeatures.ts
@@ -142,14 +142,6 @@ export class CustomFeatures {
                this.existsInList(etermaxGames, gameId);
     }
 
-    public static isAdmobTimeoutWhitelisted(gameId: string): boolean {
-        const fanateeGames = ['1344431', '1346246', '56659', '1225669'];
-        const etermaxGames = ['20721', '20723', '89611', '1781085', '1781084'];
-
-        return this.existsInList(fanateeGames, gameId) ||
-               this.existsInList(etermaxGames, gameId);
-    }
-
     public static shouldDisableBannerRefresh(gameId: string): boolean {
         if (gameId === '2962474') {
             return true;