+++ src/ts/Ads/Utilities/SDKMetrics.ts
@@ -1,6 +1,6 @@
 import { MetricInstance } from 'Ads/Networking/MetricInstance';
 
-export enum ProgrammaticTrackingError {
+export enum ErrorMetric {
     TooLargeFile = 'too_large_file', // a file 20mb and over are considered too large
     BannerRequestError = 'banner_request_error',
     AdmobTestHttpError = 'admob_video_http_error',
@@ -134,9 +134,9 @@ export enum MediationMetric {
 
 export type TimingEvent = InitializationMetric | MediationMetric;
 
-export type PTSEvent = TimingEvent | AdmobMetric | BannerMetric | CachingMetric | ChinaMetric | VastMetric | MraidMetric | MiscellaneousMetric | LoadMetric | ProgrammaticTrackingError | OMMetric | AUIMetric;
+export type PTSEvent = TimingEvent | AdmobMetric | BannerMetric | CachingMetric | ChinaMetric | VastMetric | MraidMetric | MiscellaneousMetric | LoadMetric | ErrorMetric | OMMetric | AUIMetric;
 
-export class ProgrammaticTrackingService {
+export class SDKMetrics {
 
     private static _metricInstance: MetricInstance;
 
@@ -158,20 +158,12 @@ export class ProgrammaticTrackingService {
         this._metricInstance.reportMetricEventWithTags(event, tags);
     }
 
-    public static reportErrorEvent(event: PTSEvent, adType: string, seatId?: number): void {
-        this._metricInstance.reportErrorEvent(event, adType, seatId);
-    }
-
     public static reportTimingEvent(event: TimingEvent, value: number): void {
         this._metricInstance.reportTimingEvent(event, value);
     }
 
     public static reportTimingEventWithTags(event: TimingEvent, value: number, tags: string[]): void {
         this._metricInstance.reportTimingEventWithTags(event, value, tags);
-    }
-
-    public static batchEvent(metric: TimingEvent, value: number): void {
-        this._metricInstance.batchEvent(metric, value);
     }
 
     public static sendBatchedEvents(): void {