+++ src/ts/Privacy/__mocks__/PrivacySDK.ts
@@ -1,7 +1,11 @@
 import { PrivacySDK as Base } from 'Privacy/PrivacySDK';
 
-export type PrivacySDKMock = Base & {};
+export type PrivacySDKMock = Base & {
+    isGDPREnabled: jest.Mock;
+};
 
 export const PrivacySDK = jest.fn(() => {
-    return <PrivacySDKMock>{};
+    return <PrivacySDKMock>{
+        isGDPREnabled: jest.fn()
+    };
 });